check op CLI version and show the setup dialog again if not met
